v.1.2.8 - debounced user input
it was possible to stack mutliple turns, such that in processing one turn a piece had begun to be moved. the position in the map disagreed with the position in the object. the other thread then processed the turns for the map and the state was incongruent. this could lead to errors, unresponsiveness, and otherwise bad stuff. debouncing is a simpler fix than promises, and it promotes healthy play in a way that optimally responsive design by promise may not.